Refactor: provision script for better output and logic

+++ scripts/provision_submodules.sh
@@ -26,6 +26,7 @@ done
 
 # Paths
 LIB_PATH="src/lib"
+ENV_FILE="./.env"
 
 NETWORK="colonyNetwork"
 SERVER="colonyServer"
@@ -48,14 +49,33 @@ log() {
 
 warn() {
   # Colors
+  RED=`tput setaf 3`
+  NC=`tput sgr0`
+  # Weights
+  BOLD=`tput bold`
+  echo
+  echo "${RED}${BOLD}$1${NC}"
+  echo
+}
+
+err() {
+  # Colors
   RED=`tput setaf 1`
   NC=`tput sgr0`
   # Weights
   BOLD=`tput bold`
+  echo
   echo "${RED}${BOLD}$1${NC}"
+  echo
 }
 
-cp .env.example .env
+# Setup the dapp's env file
+if [ -f "$ENV_FILE" ]; then
+    warn "The Dapp .env file already exists, skipping generating it"
+else
+    log "Generating the \"Dapp's\" submodule .env file"
+    cp .env.example .env
+fi
 
 # Update / re-pull submodules
 log "Initialize submodule libs"
@@ -69,6 +89,8 @@ then
     $YARN --pure-lockfile
     DISABLE_DOCKER=true $YARN provision:token:contracts
     cd ${ROOT_PATH}
+else
+    warn "Skipping '${NETWORK}' submodule provision"
 fi
 
 if [ "$SKIP_SERVER_BUILD" != true ]
@@ -79,6 +101,8 @@ then
     mkdir -p mongo-data
     npm install
     cd ${ROOT_PATH}
+else
+    warn "Skipping '${SERVER}' submodule provision"
 fi
 
 # Mock reputation miner
@@ -91,18 +115,21 @@ then
     log "Installing the '${ORACLE}' submodule node_modules"
     npm install
     cd ${ROOT_PATH}
+else
+    warn "Skipping '${ORACLE}' submodule provision"
 fi
 
 # Subgraph
 if [ "$SKIP_SUBGRAPH_BUILD" != true ]
 then
-    warn "If this is your first time installing, @graphprotocol/graph-ts will take a long time"
-    echo
+    err "If this is your first time installing, \"@graphprotocol/graph-ts\" will take a long time"
     log "Building the '${SUBGRAPH}' submodule"
     cd "${ROOT_PATH}/${LIB_PATH}/${SUBGRAPH}"
     log "Installing the '${SUBGRAPH}' submodule node_modules"
     npm install
     cd ${ROOT_PATH}
+else
+    warn "Skipping '${SUBGRAPH}' submodule provision"
 fi
 
 # Graph Node
@@ -113,4 +140,6 @@ then
     log "Installing the '${GRAPH_NODE}' submodule node_modules"
     npm install
     cd ${ROOT_PATH}
+else
+    warn "Skipping '${GRAPH_NODE}' submodule provision"
 fi
